move Model.js to Objects
