+++ assets/tilemaps/level02.json
@@ -3,7 +3,8 @@
     {
      "export":
         {
-         "format":"json"
+         "format":"json",
+         "target":"."
         }
     },
  "height":7,
@@ -131,7 +132,7 @@
                  "visible":true,
                  "width":64,
                  "x":704,
-                 "y":64
+                 "y":192
                 }],
          "opacity":1,
          "type":"objectgroup",
