+++ assets/tilemaps/level02.tmx
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <map version="1.4" tiledversion="1.4.1" orientation="orthogonal" renderorder="right-down" width="14" height="7" tilewidth="64" tileheight="64" infinite="0" nextlayerid="8" nextobjectid="20">
  <editorsettings>
-  <export format="json"/>
+  <export target="." format="json"/>
  </editorsettings>
  <tileset firstgid="1" name="kenney_simple_platformer" tilewidth="64" tileheight="64" tilecount="98" columns="14">
   <image source="../tilesets/platformPack_tilesheet.png" width="896" height="448"/>
@@ -27,6 +27,6 @@
   <object id="19" gid="71" x="704" y="256" width="64" height="64"/>
  </objectgroup>
  <objectgroup id="7" name="Diamonds">
-  <object id="12" gid="50" x="704" y="64" width="64" height="64"/>
+  <object id="12" gid="50" x="704" y="192" width="64" height="64"/>
  </objectgroup>
 </map>
