+++ src/index.js
@@ -6,7 +6,7 @@ import PreloaderScene from './Scenes/PreloaderScene';
 import TitleScene from './Scenes/TitleScene';
 import OptionsScene from './Scenes/OptionsScene';
 import CreditsScene from './Scenes/CreditsScene';
-import Model from './Model';
+import Model from './Objects/Model';
 
 class Game extends Phaser.Game {
   constructor () {
